Calculate_Press_correction: relative error. Returns delta_p_max Functions for Matrix: max(), diff(), Summ() added or moved to Matrix.h int -> size_t

+++ src/BiCGStab.cpp
@@ -46,8 +46,8 @@ void BiCGStab(Matrix &x, ublas::matrix<Template> &A, Matrix &b, Param par, Direc
 	r = Operator_Ax(A, x, par, Dir);
 
 
-	for (int j = 0; j < n2; ++j){
-		for (int i = 0; i < n1; ++i){
+	for (size_t j = 0; j < n2; ++j){
+		for (size_t i = 0; i < n1; ++i){
 			help_value = b[i][j] - r[i][j];
 
 			r[i][j] = help_value;
@@ -75,8 +75,8 @@ void BiCGStab(Matrix &x, ublas::matrix<Template> &A, Matrix &b, Param par, Direc
 
 		// 3.
 
-		for (int i = 0; i < n1; ++i){
-			for (int j = 0; j < n2; ++j){
+		for (size_t i = 0; i < n1; ++i){
+			for (size_t j = 0; j < n2; ++j){
 
 				p[i][j] = r[i][j] + beta * (p[i][j] - omega * v[i][j]);
 
@@ -95,8 +95,8 @@ void BiCGStab(Matrix &x, ublas::matrix<Template> &A, Matrix &b, Param par, Direc
 
 		// 6.
 
-		for (int j = 0; j < n2; ++j){
-			for (int i = 0; i < n1; ++i){
+		for (size_t j = 0; j < n2; ++j){
+			for (size_t i = 0; i < n1; ++i){
 				s[i][j] = r[i][j] - alpha * v[i][j];
 
 
@@ -116,8 +116,8 @@ void BiCGStab(Matrix &x, ublas::matrix<Template> &A, Matrix &b, Param par, Direc
 		// 9.
 		// 10.
 
-		for (int j = 0; j < n2; ++j){
-			for (int i = 0; i < n1; ++i){
+		for (size_t j = 0; j < n2; ++j){
+			for (size_t i = 0; i < n1; ++i){
 
 				// if ( i < M1 && j < M2 ){
 				// 	continue;
@@ -159,15 +159,13 @@ void BiCGStab(Matrix &x, ublas::matrix<Template> &A, Matrix &b, Param par, Direc
 
 }
 
-double ScalarOperator(Matrix &a, Matrix &b, int n1, int n2){
+double ScalarOperator(Matrix &a, Matrix &b, size_t n1, size_t n2){
 
 	double result = 0;
-	//написать сумму скал€рных произведений строк матрицы
 
-	for (int i = 0; i < n1; ++i){
-		for (int j = 0; j < n2; ++j){
+	for (size_t i = 0; i < n1; ++i){
+		for (size_t j = 0; j < n2; ++j){
 			result += a[i][j] * b[i][j];
-
 		}
 	}
